requirements/dev.txt
notebook

# extra
adjustText
geopandas
scikit-learn
scikit-misc

# doc
git+https://github.com/has2k1/plotnine-examples#egg=plotnine_examples
jupyter
nbsphinx
numpydoc>=0.9.1
sphinx>=6.1.0

# linting & testing
ruff
black>=23.1.0
pytest-cov

# release
build
twine
wheel
